added fixtures to conftest.py
